Finite States, GameController and DungeonGenerator
Started incorporating a lot of finite state machine AI systems into the game, I reorganized the code to have a Game Controller that can act as a a way to control when to make new levels and scene transitions. I broke up the code between a Map, Mesh, and Dungeon Generator. The idea is to have the map generator work out the logic of the connectivity of rooms and then have this independently connect to the mesh generator to create the mesh of the dungeon and connect to the dungeon generator to fill the rooms with enemies and objects, with the dungeon generator only concerned with "rooms" and "passages" or "doors" so that it can ignore the lower level issues.